refactor(core): 调整模块依赖和内部可见性
- 将 hutool_core::annotation 替换为 hutool_annotation 依赖
- 添加 hutool-annotation 作为 dev-dependency 到 Cargo.toml
- 导出 date 模块及其中的各类日期时间工具类
- 重新导出 map、convert、clone 等模块功能
- 调整 NumberConverter 和 TreeEntry 的字段可见性为 pub(crate)
- 统一使用 crate::convert 代替 super::convert 路径
- 更新缓存测试中的随机数生成器类型转换
- 修复克隆测试中的方法调用方式
- 移除重复的 mirror.rs 模块实现改为 re-export
- 在 JSON 测试中添加 JsonContainer 导入
- 更新 Cargo.lock 中的依赖项和版本信息

+++ crates/hutool-cache/tests/cache_parity.rs
@@ -343,7 +343,7 @@ fn lru_cache_put_test() {
         handles.push(thread::spawn(move || {
             let key: String = (0..5)
                 .map(|_| {
-                    let idx = (i.wrapping_mul(31).wrapping_add(rand::random::<usize>())) % 26;
+                    let idx = (i.wrapping_mul(31).wrapping_add(rand::random::<u64>() as usize)) % 26;
                     (b'a' + idx as u8) as char
                 })
                 .collect();
